% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_configuration.R
\name{authentication_result}
\alias{authentication_result}
\title{Authentication result for secure connection to AkadoR}
\usage{
authentication_result(
  path_database = file.path(path.expand("~"), ".appconfig", "akador",
    "database_connection.sqlite")
)
}
\arguments{
\item{path_database}{{\link[base]{character}} expected. Default values: file.path(path.expand("~"), ".appconfig", "akador", "database_connection.sqlite"). Path to file containing data base connection, default value adapts to different operating systems}
}
\description{
The purpose of the authentication_result function enables connection authentication
}
