% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_anapo_inspector}
\alias{check_anapo_inspector}
\title{Gives the inconsistencies activity position and VMS position}
\usage{
check_anapo_inspector(
  dataframe1,
  dataframe2,
  dataframe3,
  activity_crs = 4326,
  harbour_crs = 4326,
  vms_crs = 4326,
  output,
  minimum_number_vms = 20,
  threshold = 10,
  buffer_harbour = 11100
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_anapo_inspector () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_anapo_inspector () function.}

\item{dataframe3}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_anapo_inspector () function.}

\item{activity_crs}{{\link[base]{numeric}} expected. Default values: 4326. Coordinate Reference Systems for the position activity}

\item{harbour_crs}{{\link[base]{numeric}} expected. Default values: 4326. Coordinate Reference Systems for the position harbour}

\item{vms_crs}{{\link[base]{numeric}} expected. Default values: 4326. Coordinate Reference Systems for the position VMS}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{minimum_number_vms}{{\link[base]{numeric}} expected. Default values: 20. Minimum number of VMS positions required.}

\item{threshold}{{\link[base]{numeric}} expected. Default values: 10. Maximum valid distance threshold (Nautical miles) between position and nearest VMS point.}

\item{buffer_harbour}{{\link[base]{numeric}} expected. Default values: 11100. Buffer to be used for harbour, in meter}
}
\value{
The function returns a {\link[base]{character}} with output is "message", two {\link[base]{data.frame}} with output is "report" (the first without geographical location and the second with geographical location), a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_anapo_inspector function is to provide a table of data that contains an inconsistency between activity position and VMS position
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  activity_id}}
 \item{\code{  activity_date}}
 \item{\code{  activity_time}}
 \item{\code{  activity_position}}
 \item{\code{  trip_id}}
Dataframe 2:
 \item{\code{  trip_id}}
 \item{\code{  vessel_code}}
 \item{\code{  harbour_position_departure}}
 \item{\code{  harbour_position_landing}}
Dataframe 3:
 \item{\code{  vms_date}}
 \item{\code{  vms_time}}
 \item{\code{  vms_position}}
 \item{\code{  vessel_code}}
}
}
