% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_distribution_inspector}
\alias{check_distribution_inspector}
\title{Gives the inconsistencies between the weights of small and big sample fish and the sum of the small and big weights in the associated well}
\usage{
check_distribution_inspector(
  dataframe1,
  dataframe2,
  dataframe3,
  dataframe4,
  output,
  species = c("SKJ"),
  weightcategory_small = c("W-1"),
  weightcategory_big = c("W-2"),
  weightcategory_unknown = c("W-9")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_distribution_inspector () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_distribution_inspector () function.}

\item{dataframe3}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_distribution_inspector () function.}

\item{dataframe4}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_distribution_inspector () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{species}{{\link[base]{character}} expected. Default values: c("SKJ"). Vector of species categorized as small if weight category information is missing}

\item{weightcategory_small}{{\link[base]{character}} expected. Default values: c("W-1"). Vector of small weight category codes}

\item{weightcategory_big}{{\link[base]{character}} expected. Default values: c("W-2"). Vector of big weight category codes}

\item{weightcategory_unknown}{{\link[base]{character}} expected. Default values: c("W-9"). Vector of unknown weight category codes}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_distribution_inspector  function is to provide a table of data that contains an inconsistency between the small and large sample weights and the sum of the small and big weights of the associated well
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  sample_id}}
 \item{\code{  sample_well}}
 \item{\code{  trip_id}}
 \item{\code{  sample_smallsweight}}
 \item{\code{  sample_bigsweight}}
Dataframe 2:
 \item{\code{  well_id}}
 \item{\code{  well_well}}
 \item{\code{  trip_id}}
Dataframe 3:
 \item{\code{  wellactivity_id}}
 \item{\code{  well_id}}
Dataframe 4:
 \item{\code{  wellactivityspecies_id}}
 \item{\code{  wellactivity_id}}
 \item{\code{  weightcategory_code}}
 \item{\code{  specie_name}}
 \item{\code{  wellactivityspecies_weight}}
}
}
