% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_fishing_context_inspector}
\alias{check_fishing_context_inspector}
\title{Gives the inconsistencies between the school type and the association}
\usage{
check_fishing_context_inspector(
  dataframe1,
  dataframe2,
  output,
  school_type_object = c("1")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_fishing_context_inspector () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_fishing_context_inspector () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{school_type_object}{{\link[base]{character}} expected, default : c("1"). Vector of inventory of code of the object school.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_fishing_context_inspector function is to provide a table of data that contains an inconsistency with school type and the association
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  activity_id}}
 \item{\code{  schooltype_code}}
}
\itemize{
Dataframe 2:
 \item{\code{  observedsystem_id}}
 \item{\code{  activity_id}}
 \item{\code{  schooltype_code}}
}
}
