% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_landing_total_weight_inspector}
\alias{check_landing_total_weight_inspector}
\title{Gives the inconsistencies between the total weight landed during the trip for the canneries and the sum of the weights of each landing for the canneries linked to the trip}
\usage{
check_landing_total_weight_inspector(
  dataframe1,
  dataframe2,
  output,
  epsilon = 0.01
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_weighting_inspector () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_weighting_inspector () function.}

\item{output}{{\link[base]{character}} expected.Kind of expected output. You can choose between "message", "report" or "logical".}

\item{epsilon}{{\link[base]{numeric}} expected, default : 0.01. Gives the threshold at which the difference is considered too large.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_landing_total_weight_inspector function is to provide a table of data that contains an inconsistency between the sum of the weights of each landing for the canneries and the one indicated for the trip
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  trip_id}}
 \item{\code{  trip_landingtotalweight}}
Dataframe 2:
 \item{\code{  landing_id}}
 \item{\code{  landing_weight}}
 \item{\code{  trip_id}}
}
}
