% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_little_big_inspector}
\alias{check_little_big_inspector}
\title{Gives the inconsistencies between the percentage of little and big fish sampled}
\usage{
check_little_big_inspector(
  dataframe1,
  dataframe2,
  dataframe3,
  output,
  species = c("YFT", "YFT", "BET", "BET", "ALB", "ALB"),
  measuretype = c("PD1", "FL", "PD1", "FL", "PD1", "FL"),
  sizelimit = c(24, 80, 25, 77, 23.5, 78),
  size_measure_type = c("FL", "PD1"),
  threshold = 0.9
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_little_big_inspector() function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_little_big_inspector() function.}

\item{dataframe3}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_little_big_inspector() function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{species}{{\link[base]{character}} expected. Default values:  c("YFT", "YFT", "BET", "BET", "ALB", "ALB"). Vector of the species. First criterion for identifying big fish (other values are small fish)}

\item{measuretype}{{\link[base]{character}} expected. Default values: c("PD1", "FL", "PD1", "FL", "PD1", "FL"). Vector of the size measure type. Second criterion for identifying big fish (other values are small fish)}

\item{sizelimit}{{\link[base]{numeric}} expected. Default values: c(24, 80, 25, 77, 23.5, 78). Vector for defining the lower or equal limit for size measurement. Third criterion for identifying big fish (other values are small fish)}

\item{size_measure_type}{{\link[base]{character}} expected. Default values: c("FL", "PD1"). Vector with the preferred type of size measurement for small fish and then for big fish}

\item{threshold}{{\link[base]{numeric}} expected. Default values: 0.9. Threshold for percentage of small or big fish}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_little_big_inspector function is to provide a table of data that contains an inconsistency between the percentage of little and big fish sampled
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  sample_id}}
 \item{\code{  sample_smallsweight}}
 \item{\code{  sample_bigsweight}}
 \item{\code{  sample_totalweight}}
Dataframe 2:
 \item{\code{  samplespecies_id}}
 \item{\code{  species_fao_code}}
 \item{\code{  sizemeasuretype_code}}
 \item{\code{  sample_id}}
Dataframe 3:
 \item{\code{  samplespeciesmeasure_id}}
 \item{\code{  samplespeciesmeasure_sizeclass}}
 \item{\code{  samplespeciesmeasure_count}}
 \item{\code{  samplespecies_id}}
}
}
