% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_operation_inspector}
\alias{check_operation_inspector}
\title{Gives the inconsistencies between fishing success status, vessel activity, type of school or weight caught}
\usage{
check_operation_inspector(
  dataframe1,
  output,
  vessel_activity = c("6"),
  school_type = c("0"),
  success_status_school_type = c("0", "1"),
  success_status_weight = c("0")
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_operation_inspector () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{vessel_activity}{{\link[base]{character}} expected. Default values: c("6"). Vector of inventory of codes for activities that must have a success status}

\item{school_type}{{\link[base]{character}} expected, default : c("0"). Vector of inventory of code of the school type must not have specific success status (defined success_status_school_type)}

\item{success_status_school_type}{{\link[base]{character}} expected, default : c("0", "1"). Vector of inventory of code of the success status must not have specific school type (defined school_type)}

\item{success_status_weight}{{\link[base]{character}} expected, default : c("0"). Vector of inventory of code of the success status must not have weigth}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_operation_inspector function is to provide a table of data that contains an inconsistency with succes status and vessel activity, the type of school or the weight caught
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  activity_id}}
 \item{\code{  schooltype_code}}
 \item{\code{  successstatus_code}}
 \item{\code{  activity_weight}}
 \item{\code{  vesselactivity_code}}
}
}
\examples{
dataframe1 <- data.frame(activity_id = c("1", "2", "3"),
                         schooltype_code = c("2", "1", "2"),
                         successstatus_code = c("1", "0", "2"),
                         activity_weight = c(10, 0, 0),
                         vesselactivity_code = c("6", "6", "6"))
check_operation_inspector(dataframe1, output = "report")
}
