% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_position_inspector}
\alias{check_position_inspector}
\title{Gives the inconsistencies between the ocean declared for the trip and the position of the activity}
\usage{
check_position_inspector(
  dataframe1,
  output,
  ocean_name_nonpriority = "Atlantic"
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_weighting_inspector () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{ocean_name_nonpriority}{{\link[base]{character}} expected. Default values: Atlantic. Name of the priority ocean when the point is on the border between two oceans.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", two {\link[base]{data.frame}} with output is "report" (the first without geographical location and the second with geographical location), a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_position_inspector function is to provide a table of data that contains an inconsistency with ocean declaration and activity position
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  activity_id}}
 \item{\code{  ocean_name}}
 \item{\code{  zfao_ocean}}
 \item{\code{  activity_position}}
 \item{\code{  activity_crs}}
 \item{\code{  harbour_id}}
}
}
