% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_super_sample_number_consistent_inspector}
\alias{check_super_sample_number_consistent_inspector}
\title{Gives the inconsistencies between the sample and the subsample number}
\usage{
check_super_sample_number_consistent_inspector(dataframe1, dataframe2, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_super_sample_number_consistent_inspector() function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_super_sample_number_consistent_inspector() function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_super_sample_number_consistent_inspector function is to provide a table of data that contains an inconsistency between the sample and the subsample number
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  sample_id}}
 \item{\code{  sample_supersample}}
Dataframe 2:
 \item{\code{  samplespecies_id}}
 \item{\code{  samplespecies_subsamplenumber}}
 \item{\code{  sample_id}}
}
}
