% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_temporal_limit_inspector}
\alias{check_temporal_limit_inspector}
\title{Gives the inconsistencies between trip start and end date and the dates of activity}
\usage{
check_temporal_limit_inspector(dataframe1, dataframe2, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_temporal_limit_inspector () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_temporal_limit_inspector () function.}

\item{output}{{\link[base]{character}} expected.Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", two {\link[base]{data.frame}} with output is "report" (the first at the trip level and the second at the activity date level), a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_temporal_limit_inspector function is to provide a table of data that contains an inconsistency between trip start and end date and the dates of activity (activity date outside the trip ranges, several occurrences of the activity date, ...)
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  trip_id}}
 \item{\code{  trip_startdate}}
 \item{\code{  trip_enddate}}
}
\itemize{
Dataframe 2:
 \item{\code{  route_id}}
 \item{\code{  activity_date}}
 \item{\code{  trip_id}}
}
}
