% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_time_route_inspector}
\alias{check_time_route_inspector}
\title{Gives the inconsistencies between the fishing times or sea times indicated for the route and activities carried out}
\usage{
check_time_route_inspector(
  dataframe1,
  dataframe2,
  dataframe3,
  output,
  threshold_sea_time = 24,
  threshold_fishing_time = 13,
  vessel_activity_sea_time = c("1", "2", "3", "4", "6", "6", "8", "9", "10", "11", "12",
    "13", "13", "13", "13", "13", "13", "14", "15", "17", "18", "19", "20", "22", "23",
    "24", "25", "26", "27", "29", "30", "31", "32", "36", "37", "38", "39", "50", "99",
    "101", "101", "102", "102", "103"),
  object_operation_sea_time = c(NA, NA, NA, NA, "99", NA, NA, NA, NA, NA, NA, "1", "2",
    "4", "6", "8", "9", NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA,
    NA, NA, NA, NA, NA, NA, "99", NA, "99", NA, NA),
  vessel_activity_fishing_time = c("2", "3", "4", "6", "6", "8", "12", "13", "13", "13",
    "13", "13", "13", "14", "15", "17", "19", "20", "23", "25", "26", "27", "29", "30",
    "31", "102", "102"),
  object_operation_fishing_time = c(NA, NA, NA, "99", NA, NA, NA, "1", "2", "4", "6",
    "8", "9", NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "99", NA)
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_time_route_inspector () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_time_route_inspector () function.}

\item{dataframe3}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_time_route_inspector () function.}

\item{output}{{\link[base]{character}} expected.Kind of expected output. You can choose between "message", "report" or "logical".}

\item{threshold_sea_time}{{\link[base]{numeric}} expected. Default values: 24. Maximum valid time for sea time}

\item{threshold_fishing_time}{{\link[base]{numeric}} expected. Default values: 13. Maximum valid time for fishing time}

\item{vessel_activity_sea_time}{{\link[base]{character}} expected. Default values: c("1", "2", "3", "4", "6", "6", "8", "9", "10", "11", "12", "13", "13", "13", "13", "13", "13", "14", "15", "17", "18", "19", "20", "22", "23", "24", "25", "26", "27", "29", "30", "31", "32", "36", "37", "38", "39", "50", "99", "101", "101", "102", "102", "103"). Code activities. First criterion for identifying activities that must have sea time}

\item{object_operation_sea_time}{{\link[base]{character}} expected. Default values: c(NA, NA, NA, NA, "99", NA, NA, NA, NA, NA, NA, "1", "2", "4", "6", "8", "9", NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "99", NA, "99", NA, NA). Code object operation. Second criterion for identifying activities that must have sea time (Indicate NA if you want the code object operation to be missing)}

\item{vessel_activity_fishing_time}{{\link[base]{character}} expected. Default values: c("2", "3", "4", "6", '6', "8", "12", "13", "13", "13", "13", "13", "13", "14", "15", "17", "19", "20", "23", "25", "26", "27", "29", "30", "31", "102", "102"). Code activities. First criterion for identifying activities that must have fishing time}

\item{object_operation_fishing_time}{{\link[base]{character}} expected. Default values: c(NA, NA, NA, "99", NA, NA, NA, "1", "2", "4", "6", "8", "9", NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, NA, "99", NA). Code object operation. Second criterion for identifying activities that must have fishing time (Indicate NA if you want the code object operation to be missing)}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_time_route_inspector function is to provide a table of data that contains an inconsistency between the fishing times or sea times indicated for the route and activities carried out
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  route_id}}
 \item{\code{  route_seatime}}
 \item{\code{  route_fishingtime}}
}
\itemize{
Dataframe 2:
 \item{\code{  activity_id}}
 \item{\code{  vesselactivity_code}}
 \item{\code{  route_id}}
}
\itemize{
Dataframe 3:
 \item{\code{  floatingobject_id}}
 \item{\code{  objectoperation_code}}
 \item{\code{  activity_id}}
}
}
