% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_weight_inspector}
\alias{check_weight_inspector}
\title{Gives the inconsistencies between the sum of the weight indicated for catches and the one indicated for the activity}
\usage{
check_weight_inspector(dataframe1, dataframe2, output, epsilon = 0.01)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_weight_inspector () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_weight_inspector () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{epsilon}{{\link[base]{numeric}} expected, default : 0.01. Gives the threshold at which the difference is considered too large.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_weight_inspector function is to provide a table of data that contains an inconsistency between the sum of the weight indicated for the catch and the one indicated for the activity
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  activity_id}}
 \item{\code{  activity_weight}}
}
\itemize{
Dataframe 2:
 \item{\code{  catch_id}}
 \item{\code{  catch_weight}}
 \item{\code{  activity_id}}
}
}
