% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_weighting_sample_inspector}
\alias{check_weighting_sample_inspector}
\title{Gives the inconsistencies between the sample weighting and catch weight}
\usage{
check_weighting_sample_inspector(
  dataframe1,
  dataframe2,
  output,
  species = c("YFT", "SKJ", "BET", "ALB", "LTA", "FRI", "TUN", "KAW", "LOT"),
  speciesfate = "6",
  epsilon = 0.01
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_weighting_sample_inspector () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_weighting_sample_inspector () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{species}{{\link[base]{character}} expected. Default values: c("YFT", "SKJ", "BET", "ALB", "LTA", "FRI", "TUN", "KAW", "LOT"). list of the inventory of species (FAO code) used to compare to sample weighting.}

\item{speciesfate}{{\link[base]{character}} expected. Default values: "6". Vector of inventory of fate used to compare to sample weighting.}

\item{epsilon}{{\link[base]{numeric}} expected, default : 0.01. Gives the threshold at which the difference is considered too large.}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_weighting_sample_inspector function is to provide a table of data that contains an inconsistency between the sample weighting and catch weight for activity
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  catch_id}}
 \item{\code{  catch_weight}}
 \item{\code{  speciesfate_code}}
 \item{\code{  species_fao_code}}
 \item{\code{  activity_id}}
Dataframe 2:
 \item{\code{  sample_id}}
 \item{\code{  sampleactivity_weightedweight}}
 \item{\code{  activity_id}}
}
}
