% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_well_number_consistent_inspector}
\alias{check_well_number_consistent_inspector}
\title{Gives the inconsistencies between sample well number and associated trip well numbers}
\usage{
check_well_number_consistent_inspector(dataframe1, dataframe2, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_well_number_consistent_inspector () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_well_number_consistent_inspector () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_well_number_consistent_inspector  function is to provide a table of data that contains an inconsistency between sample well number and associated trip well numbers
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  sample_id}}
 \item{\code{  sample_well}}
 \item{\code{  trip_id}}
Dataframe 2:
 \item{\code{  trip_id}}
 \item{\code{  well_well}}
}
}
