% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_toolbox.R
\name{column_grounding}
\alias{column_grounding}
\title{Function detects activity linked solely to grounding}
\usage{
column_grounding(
  data,
  data_transmittingbuoy,
  transmittingbuoyoperation_grounding_code = c("4", "5")
)
}
\arguments{
\item{data}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the column_grounding () function.}

\item{data_transmittingbuoy}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the column_grounding () function.}

\item{transmittingbuoyoperation_grounding_code}{{\link[base]{character}} expected. Default values: c("4", "5"). vector of buoy operation codes indicating object grounding}
}
\value{
The function returns a {\link[base]{data.frame}}
}
\description{
Function detects activity linked solely to grounding
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
data :
 \item{\code{  activity_id}}
}
\itemize{
data_transmittingbuoy :
 \item{\code{  transmittingbuoyoperation_code}}
 \item{\code{  activity_id}}
}
}
\examples{
data <- data.frame(activity_id = c("1", "2", "3"))
data_transmittingbuoy <- data.frame(transmittingbuoyoperation_code = c("1", "4"),
                                    activity_id = c("1", "3"))
column_grounding(data, data_transmittingbuoy)
}
