% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_toolbox.R
\name{coordinate_dd_to_dmd}
\alias{coordinate_dd_to_dmd}
\title{Converting DD coordinates in DDM}
\usage{
coordinate_dd_to_dmd(coordinate, latitude)
}
\arguments{
\item{coordinate}{{\link[base]{numeric}} expected. Vector of coordinate values}

\item{latitude}{{\link[base]{logical}} expected. Logical which indicates whether the values correspond to the latitude (TRUE) of the coordinates or to the longitude (FALSE).}
}
\value{
The function returns a {\link[base]{character}}
}
\description{
Function for converting DD (Decimal Degrees) coordinates in DDM (Degrees, Decimal Minutes)
}
\examples{
coordinate_dd_to_dmd(coordinate = c(0.1, 5.69), latitude = TRUE)
}
