% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_toolbox.R
\name{data_to_list}
\alias{data_to_list}
\title{Function to create a data.frame in list}
\usage{
data_to_list(
  data,
  name_col_dataplot,
  colname_id,
  colname_plot,
  colname_info,
  rename_colname_info
)
}
\arguments{
\item{data}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the data_to_list () function.}

\item{name_col_dataplot}{{\link[base]{character}} expected. Name of the sub-list containing the table}

\item{colname_id}{{\link[base]{character}} expected. Name of the column in data used to identify a group}

\item{colname_plot}{{\link[base]{character}} expected. Vector of column names in data to be grouped in an array because it contains several values per group}

\item{colname_info}{{\link[base]{character}} expected. Vector of column names in data to be organized in separate sub-lists because it contains a single value per group}

\item{rename_colname_info}{{\link[base]{character}} expected. Vector with sub-list names to be renamed}
}
\value{
The function returns a {\link[base]{list}} with the sub list
}
\description{
Function to create a data.frame in list
}
\examples{
dataframe1 <- data.frame(trip_id = c("1", "1", "2", "2", "2"),
                         trip_startdate = as.Date(c("2020/01/01", "2020/01/01", "2020/01/01",
                                                    "2020/01/01", "2020/01/01")),
                         trip_enddate = as.Date(c("2020/01/02", "2020/01/02", "2020/01/02",
                                                  "2020/01/02", "2020/01/02")),
                         activity_date = as.Date(c("2020/01/01", "2020/01/02","2020/01/01",
                                                   "2020/01/02", "2020/01/03")),
                         logical = c(TRUE, TRUE, TRUE, TRUE, FALSE))
data_to_list(data = dataframe1, name_col_dataplot = "data", colname_id = "trip_id",
             colname_plot = c("activity_date", "logical"),
             colname_info = c("trip_startdate", "trip_enddate"),
             rename_colname_info = c("startdate", "enddate"))
}
