% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_specific_check.R
\name{display_anapo}
\alias{display_anapo}
\title{Function to change table display anapo}
\usage{
display_anapo(dataframe1, dataframe2, dataframe3, dataframe4)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or the second output of the function {\link[codama]{check_anapo_inspector}}, which must be done before using the display_anapo () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the display_anapo () function.}

\item{dataframe3}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the display_anapo () function.}

\item{dataframe4}{{\link[base]{data.frame}} expected. Csv or the first output of the function {\link[codama]{check_anapo_inspector}}, which must be done before using the display_anapo () function.}
}
\value{
The function returns a {\link[base]{data.frame}}
}
\description{
Adds information about groundings and previous and next activities
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  activity_id}}
 \item{\code{  activity_date}}
 \item{\code{  activity_time}}
 \item{\code{  activity_position}}
}
\itemize{
Dataframe 2:
 \item{\code{  trip_id}}
 \item{\code{  activity_id}}
 \item{\code{  activity_number}}
}
\itemize{
Dataframe 3:
 \item{\code{  transmittingbuoy_id}}
 \item{\code{  transmittingbuoyoperation_code}}
 \item{\code{  activity_id}}
}
\itemize{
Dataframe 4:
 \item{\code{  activity_id}}
 \item{\code{  logical}}
 \item{\code{  nb_vms}}
 \item{\code{  min_distance}}
 \item{\code{  max_score}}
}
}
\examples{
dataframe1 <- data.frame(activity_id  = c("2", "4", "6"),
                         activity_date = as.Date(c("2020/01/12", "2020/01/13", "2020/01/13")),
                         activity_time = c("05:26:01", "03:12:34", "23:26:47"),
                         activity_position = c("POINT (0 0)", "POINT (4 4)", "POINT (3 0.6)"))
dataframe2 <- data.frame(trip_id  = c("1", "1", "1"),
                         activity_id = c("2", "4", "6"),
                         activity_number = c(1L, 1L, 2L))
dataframe3 <- data.frame(transmittingbuoy_id  = c("1"),
                         transmittingbuoyoperation_code = c("4"),
                         activity_id = c("6"))
dataframe4 <- data.frame(activity_id  = c("2", "4", "6"),
                         logical = c(TRUE, FALSE, FALSE),
                         nb_vms = c(1L, NA, 2L),
                         min_distance = c(6.004055, 84.717419, 18.012165),
                         max_score = c(NA, 0.0000000, 0.2094411))
units::install_unit("NM", "1852 m", "Nautical mile")
dataframe4$min_distance <- units::set_units(dataframe4$min_distance, NM)
display_anapo(dataframe1, dataframe2, dataframe3, dataframe4)
}
