% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_specific_check.R
\name{display_little_big}
\alias{display_little_big}
\title{Function to change table display little_big}
\usage{
display_little_big(dataframe1)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or the first output of the function {\link[AkadoR]{check_little_big_inspector}}, which must be done before using the display_little_big () function.}
}
\value{
The function returns a {\link[base]{data.frame}} with the modified percentage
}
\description{
Changes the number of decimal places in percentage
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  little_percentage}}
 \item{\code{  big_percentage}}
 \item{\code{  measure1_percentage}}
 \item{\code{  measure2_percentage}}
}
}
\examples{
dataframe1 <- data.frame(little_percentage = c(0.0001, 0.19),
                         big_percentage = c(0.1999, 0.61),
                         measure1_percentage = c(0.5, 0.005),
                         measure2_percentage = c(0.3, 0.195))
display_little_big(dataframe1)
}
