% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_specific_check.R
\name{display_raising_factor}
\alias{display_raising_factor}
\title{Function to change table display raising_factor}
\usage{
display_raising_factor(dataframe1)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or the first output of the function {\link[codama]{check_raising_factor_inspector}}, which must be done before using the display_raising_factor () function.}
}
\value{
The function returns a {\link[base]{data.frame}} with the modified RF1
}
\description{
Changes the number of decimal places in RF1
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  rf1}}
}
}
\examples{
dataframe1 <- data.frame(rf1 = c(1, 1.11111111111111, 2.55555))
display_raising_factor(dataframe1)
}
