% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_radiobuttons_type_check.R
\name{mod_radiobuttons_type_check_server}
\alias{mod_radiobuttons_type_check_server}
\title{#' @title radiobuttons_type_check Server Functions}
\usage{
mod_radiobuttons_type_check_server(id, type_check_info)
}
\arguments{
\item{id}{Internal parameters for shiny}

\item{type_check_info}{{\link[base]{list}} expected. Information about the dynamic type of check}
}
\value{
The function returns nothing, instantiating the radiobuttons
}
\description{
A shiny Module for creation of all tab
}
\details{
\itemize{
type_check_info:
\item{\code{  title : Button title, mandatory, character expected}}
\item{\code{  id : Button selection identification, mandatory, character expected}}
\item{\code{  text : Display text for button selection, mandatory, character expected}}
}
}
