% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_radiobuttons_type_check.R
\name{mod_radiobuttons_type_check_ui}
\alias{mod_radiobuttons_type_check_ui}
\title{radiobuttons_type_check UI Function}
\usage{
mod_radiobuttons_type_check_ui(id)
}
\arguments{
\item{id}{Internal parameters for shiny}
}
\value{
The function returns a {\link[shiny]{tagList}} with a {\link[shiny]{radioButtons}}
}
\description{
A shiny Module for creation of radio button to select the display of controls
}
