% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_content.R
\name{mod_tab_content_ui}
\alias{mod_tab_content_ui}
\title{tab_content UI Function}
\usage{
mod_tab_content_ui(id)
}
\arguments{
\item{id}{Internal parameters for shiny}
}
\value{
The function returns a {\link[shiny]{tagList}} with a {\link[shinydashboard]{tabItems}} and {\link[shinydashboard]{tabItem}}
}
\description{
A shiny Module for creation of all tab
}
