% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_menu.R
\name{mod_tab_menu_server}
\alias{mod_tab_menu_server}
\title{tab_menu Server Functions}
\usage{
mod_tab_menu_server(id, tab_info)
}
\arguments{
\item{id}{Internal parameters for shiny}

\item{tab_info}{{\link[base]{list}} expected. Information about the dynamic tab display}
}
\value{
The function returns a reactive valu
}
\description{
A shiny Module for creation of all menus
}
\details{
\itemize{
tab_info:
 \item{\code{  id : Tab identification, mandatory, character expected}}
 \item{\code{  title : Name to be displayed for tab, mandatory, character expected}}
 \item{\code{  icon : Picture to be displayed for tab, optional, character expected}}
}
}
