% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_menu.R
\name{mod_tab_menu_ui}
\alias{mod_tab_menu_ui}
\title{tab_menu UI Function}
\usage{
mod_tab_menu_ui(id)
}
\arguments{
\item{id}{Internal parameters for shiny}
}
\value{
The function returns a {\link[shiny]{tagList}} with a {\link[shinydashboard]{sidebarMenu}} and {\link[shinydashboard]{menuItem}}
}
\description{
A shiny Module for creation of all menus
}
