% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{mod_table_server}
\alias{mod_table_server}
\title{table Server Functions}
\usage{
mod_table_server(id, data, name, type_line_check, column_no_wrap = NULL)
}
\arguments{
\item{id}{Internal parameters for shiny}

\item{data}{{\link[base]{list}} expected. Reactive list containing results tables for all checks}

\item{name}{{\link[base]{character}} expected. Control identifier in data to be displayed}

\item{type_line_check}{{\link[base]{character}} expected. Reactive value containing the type of line selected by the user}

\item{column_no_wrap}{{\link[base]{integer}} expected. Default values: NULL. Column numbers that should not be subject to automatic line breaks}
}
\value{
The function returns nothing, instantiating the table
}
\description{
A shiny Module for creation of all table
}
