% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{mod_table_ui}
\alias{mod_table_ui}
\title{table UI Function}
\usage{
mod_table_ui(
  id,
  title = NULL,
  size_box = "col-sm-12 col-md-6 col-lg-6",
  text = NULL
)
}
\arguments{
\item{id}{Internal parameters for shiny}

\item{title}{{\link[base]{character}} expected. Default values: NULL. Name to be displayed for check}

\item{size_box}{{\link[base]{character}} expected. Default values: "col-sm-12 col-md-6 col-lg-6". Check size specification for each window size type (format "col-sm-your_size col-md-your_size col-lg-your_size")}

\item{text}{{\link[base]{character}} expected. Default values: NULL. Text to be displayed for check}
}
\value{
The function returns a {\link[shiny]{tagList}} with a {\link[shinydashboard]{box}} and {\link[DT]{DTOutput}}
}
\description{
A shiny Module for creation of all table
}
