% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_specific_check.R
\name{plot_anapo_activity}
\alias{plot_anapo_activity}
\title{Function to create the plot of the consistency of the position for VMS}
\usage{
plot_anapo_activity(data_vms, crs_vms, date_vms)
}
\arguments{
\item{data_vms}{{\link[base]{data.frame}} expected. Csv or an element of the output of function {\link[AkadoR]{plot_anapo_activity_data}}, which must be done before using the plot_anapo_activity () function.}

\item{crs_vms}{{\link[base]{numeric}} expected. Coordinates Reference System for position VMS}

\item{date_vms}{{\link[base]{Date}} expected. VMS date}
}
\value{
The function returns a {\link[plotly]{plotly}}
# Example 1
data_vms <- data.frame(vms_time = c("10:55:15", "22:32:17"),
                       vms_position = c("POINT (0 0.1)", "POINT (3 0.3)"))
plot_anapo_activity(data_vms, 4326, as.Date(c("2020/01/13")))
}
\description{
Function to create the plot
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
data_vms :
 \item{\code{  vms_time}}
 \item{\code{  vms_position}}
}
}
