% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_specific_check.R
\name{plot_anapo_activity_data}
\alias{plot_anapo_activity_data}
\title{Function to create list data/argument for the plot plot_anapo_activity}
\usage{
plot_anapo_activity_data(dataframe1, dataframe2)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or the second output of the function {\link[AkadoR]{check_anapo_activity_consistent_inspector}}, which must be done before using the plot_anapo_activity_data () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the plot_anapo_activity_data () function.}
}
\value{
The function returns a {\link[base]{list}} with the sub list is a data for one plot
}
\description{
Function to create list data/argument
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  vms_id}}
 \item{\code{  vessel_code}}
 \item{\code{  vms_date}}
 \item{\code{  vessel_type}}
}
\itemize{
Dataframe 2:
 \item{\code{  vms_id}}
 \item{\code{  vms_time}}
 \item{\code{  vms_position}}
 \item{\code{  vms_crs}}
}
}
\examples{
dataframe1 <- data.frame(vms_id = c("1", "2"),
                         vessel_code = c("0", "0"),
                         vms_date = as.Date(c("2020/01/13", "2020/01/13")),
                         vessel_type = c("6", "6"))
dataframe2 <- data.frame(vms_id = c("1", "2"),
                         vms_time = c("10:55:15", "22:32:17"),
                         vms_position = c("POINT (0 0.1)", "POINT (3 0.3)"),
                         vms_crs = c(4326, 4326))
plot_anapo_activity_data(dataframe1, dataframe2)
}
