% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_specific_check.R
\name{plot_anapo_activity_windows}
\alias{plot_anapo_activity_windows}
\title{Function to create the text for windows with plot anapo activity}
\usage{
plot_anapo_activity_windows(vessel_code, date_vms, vessel_type)
}
\arguments{
\item{vessel_code}{{\link[base]{character}} expected. Vessel code}

\item{date_vms}{{\link[base]{Date}} expected. VMS date}

\item{vessel_type}{{\link[base]{character}} expected. Vessel type}
}
\value{
The function returns a {\link[base]{character}}
}
\description{
Function to create the windows with plot of the consistency of the position for VMS
}
\examples{
plot_anapo_activity_windows("0", as.Date(c("2020/01/13")), "6")
}
