% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_specific_check.R
\name{plot_anapo_windows}
\alias{plot_anapo_windows}
\title{Function to create the text for windows with plot anapo}
\usage{
plot_anapo_windows(data_activity)
}
\arguments{
\item{data_activity}{{\link[base]{data.frame}} expected. Csv or an element of the output of function {\link[AkadoR]{plot_anapo_data}}, which must be done before using the plot_anapo () function.}
}
\value{
The function returns a {\link[base]{character}}
}
\description{
Function to create the windows with plot of the consistency of the position for the activity and VMS
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
data_activity :
 \item{\code{  vessel_code}}
 \item{\code{  trip_enddate}}
 \item{\code{  activity_date}}
 \item{\code{  activity_time}}
 \item{\code{  activity_number}}
 \item{\code{  vesselactivity_code}}
 \item{\code{  X}}
 \item{\code{  Y}}
 \item{\code{  grounding}}
}
}
\examples{
data_activity <- data.frame(vessel_code = c("0"),
                            trip_enddate = as.Date(c("2020/01/20")),
                            activity_date =  as.Date(c("2020/01/12")),
                            activity_time = c("05:26:01"),
                            activity_number = 1L,
                            vesselactivity_code = "2",
                            X = c("E0°0'"),
                            Y = c("N0°0'"),
                            grounding = FALSE)
plot_anapo_windows(data_activity)
}
