% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_specific_check.R
\name{plot_eez_data}
\alias{plot_eez_data}
\title{Function to create list data/argument for the plot eez}
\usage{
plot_eez_data(dataframe1, dataframe2)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or the second output of the function {\link[AkadoR]{check_eez_inspector}}, which must be done before using the plot_eez_data () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the plot_eez_data () function.}
}
\value{
The function returns a {\link[base]{list}} with the sub list is a data for one plot
}
\description{
Function to create list data/argument
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  activity_id}}
 \item{\code{  activity_position}}
 \item{\code{  activity_crs}}
 \item{\code{  fpazone_code}}
 \item{\code{  fpazone_country_iso3}}
 \item{\code{  eez_calculated}}
}
\itemize{
Dataframe 2:
 \item{\code{  activity_id}}
 \item{\code{  vessel_code}}
 \item{\code{  trip_enddate}}
 \item{\code{  activity_date}}
 \item{\code{  activity_number}}
}
}
\examples{
dataframe1 <- data.frame(activity_id = c("1", "2"),
                         activity_position = c("POINT (0.1 0)", "POINT (0 0)"),
                         fpazone_code = c("SYC", "XSG"),
                         fpazone_country_iso3 = c("SYC", "XXX"),
                         eez_calculated = c("SYC", "XSG"),
                         activity_crs = c(4326, 4326))
dataframe2 <- data.frame(activity_id = c("1", "2"),
                         vessel_code = c("0", "0"),
                         trip_enddate = as.Date(c("2020/01/10", "2020/01/10")),
                         activity_date = as.Date(c("2020/01/09", "2020/01/10")),
                         activity_number = c(1L, 1L))
plot_eez_data(dataframe1, dataframe2)
}
