% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_specific_check.R
\name{plot_eez_windows}
\alias{plot_eez_windows}
\title{Function to create the text for windows with plot eez}
\usage{
plot_eez_windows(
  vessel_code,
  trip_enddate,
  activity_date,
  activity_number,
  X,
  Y,
  fpazone_code,
  fpazone_country_iso3,
  eez_calculated
)
}
\arguments{
\item{vessel_code}{{\link[base]{character}} expected. Vessel code}

\item{trip_enddate}{{\link[base]{Date}} expected. Trip end date}

\item{activity_date}{{\link[base]{Date}} expected. Activity date}

\item{activity_number}{{\link[base]{integer}} expected. Activity number}

\item{X}{{\link[base]{character}} expected. Position longitude}

\item{Y}{{\link[base]{character}} expected. Position latitude}

\item{fpazone_code}{{\link[base]{character}} expected. Fpa zone code}

\item{fpazone_country_iso3}{{\link[base]{character}} expected. Iso3 country code for fpa zone}

\item{eez_calculated}{{\link[base]{character}} expected. Calculated zee zone code}
}
\value{
The function returns a {\link[base]{character}}
}
\description{
Function to create the windows with plot of the consistency of the eez for the activity
}
\examples{
plot_eez_windows(vessel_code = "0", trip_enddate = as.Date("2020/01/10"),
                 activity_date = as.Date("2020/01/09"), activity_number = 1L, X = "E0°6'",
                 Y = "N0°0'", fpazone_code = "SYC", fpazone_country_iso3 = "SYC",
                 eez_calculated = "SYC")
}
