% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_specific_check.R
\name{plot_position}
\alias{plot_position}
\title{Function to create the plot of the consistency of the position for the activity}
\usage{
plot_position(data)
}
\arguments{
\item{data}{{\link[base]{data.frame}} expected. Csv or an element of the output of function {\link[AkadoR]{plot_position_data}}, which must be done before using the plot_position () function.}
}
\value{
The function returns a {\link[plotly]{plotly}}
# Example 1
data <- data.frame(activity_position = "POINT (0 0)",
                   activity_crs = 4326)
plot_position(data)
}
\description{
Function to create the plot
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
data :
 \item{\code{  activity_position}}
 \item{\code{  activity_crs}}
}
}
