% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_specific_check.R
\name{plot_position_windows}
\alias{plot_position_windows}
\title{Function to create the text for windows with plot position}
\usage{
plot_position_windows(
  vessel_code,
  trip_enddate,
  activity_date,
  activity_number,
  X,
  Y,
  type,
  ocean_label,
  ocean_calculate
)
}
\arguments{
\item{vessel_code}{{\link[base]{character}} expected. Vessel code}

\item{trip_enddate}{{\link[base]{Date}} expected. Trip end date}

\item{activity_date}{{\link[base]{Date}} expected. Activity date}

\item{activity_number}{{\link[base]{integer}} expected. Activity number}

\item{X}{{\link[base]{character}} expected. Position longitude}

\item{Y}{{\link[base]{character}} expected. Position latitude}

\item{type}{{\link[base]{character}} expected. Type of inconsistency}

\item{ocean_label}{{\link[base]{character}} expected. Ocean label}

\item{ocean_calculate}{{\link[base]{character}} expected. Ocean calculate}
}
\value{
The function returns a {\link[base]{character}}
}
\description{
Function to create the windows with plot of the consistency of the position for the activity
}
\examples{
plot_position_windows(vessel_code = "0", trip_enddate = as.Date("2020/01/10"),
                      activity_date = as.Date("2020/01/09"), activity_number = 1L, X = "E0°0'",
                      Y = "N0°0'", type = "Sea", ocean_label = "ocean_1",
                      ocean_calculate = "ocean_1")
}
