% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_specific_check.R
\name{plot_temporal_limit}
\alias{plot_temporal_limit}
\title{Function to create the plot of the consistency of the dates by trip}
\usage{
plot_temporal_limit(data, startdate, enddate)
}
\arguments{
\item{data}{{\link[base]{data.frame}} expected. Csv or an element of the output of function {\link[AkadoR]{plot_temporal_limit_data}}, which must be done before using the plot_temporal_limit () function.}

\item{startdate}{{\link[base]{Date}} expected. Trip start date}

\item{enddate}{{\link[base]{Date}} expected. Trip end date}
}
\value{
The function returns a {\link[plotly]{plotly}}
# Example 1
data_1 <- data.frame(activity_date = as.Date(c("2020/01/01", "2020/01/02")),
                   logical = c(TRUE, TRUE),
                   count_freq = c(1L, 1L))
plot_temporal_limit(data_1, as.Date("2020/01/01"), as.Date("2020/01/02"))
# Example 2
data_2 <- data.frame(activity_date = as.Date(c("2020/01/01", "2020/01/02", "2020/01/03")),
                   logical = c(TRUE, TRUE, FALSE),
                   count_freq = c(1L, 1L, 1L))
plot_temporal_limit(data_2, as.Date("2020/01/01"), as.Date("2020/01/02"))
}
\description{
Function to create the plot
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
data :
 \item{\code{  activity_date}}
 \item{\code{  count_freq}}
 \item{\code{  logical}}
}
}
