% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_specific_check.R
\name{plot_temporal_limit_windows}
\alias{plot_temporal_limit_windows}
\title{Function to create the text for windows with plot temporal}
\usage{
plot_temporal_limit_windows(vessel_code, enddate)
}
\arguments{
\item{vessel_code}{{\link[base]{character}} expected. Vessel code}

\item{enddate}{{\link[base]{Date}} expected. Trip end date}
}
\value{
The function returns a {\link[base]{character}}
}
\description{
Function to create the text for windows with plot of the consistency of the dates by trip
}
\examples{
plot_temporal_limit_windows("0", as.Date("2020/01/02"))
}
