% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_configuration.R
\name{set_authentication_result}
\alias{set_authentication_result}
\title{AkadoR set authentication}
\usage{
set_authentication_result(
  path_file_configuration = file.path(path.expand("~"), ".appconfig", "akador",
    "configuration_file.yml")
)
}
\arguments{
\item{path_file_configuration}{{\link[base]{character}} expected. Default values: file.path(path.expand("~"), ".appconfig", "akador", "configuration_file.yml"). Path to file containing configuration parameters, default value adapts to different operating systems}
}
\description{
The purpose of the set_authentication_result function allows application launch parameters to be modified by an external file
}
