% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{table_display}
\alias{table_display}
\title{Transforms table display}
\usage{
table_display(data, data_info, type_inconsistency, rename, order)
}
\arguments{
\item{data}{{\link[base]{data.frame}} expected. data frame containing check information}

\item{data_info}{{\link[base]{data.frame}} expected. data frame containing information to enable the user to identify lines}

\item{type_inconsistency}{{\link[base]{character}} expected. Control type. You can choose between "error", "warning" or "info".}

\item{rename}{{\link[base]{list}} expected. List of column names to be renamed (the left part is the column names, the right part is the new names)}

\item{order}{{\link[base]{character}} expected. Column names according to which rows are to be sorted.}
}
\value{
The function returns a {\link[base]{data.frame}}
}
\description{
Shiny function : Function that formats the display of check results, positions, row order and column names
}
