% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_toolbox.R
\name{text_object_more_or_less}
\alias{text_object_more_or_less}
\title{Function to list elements with a number of occurrences other than 2}
\usage{
text_object_more_or_less(id, result_check)
}
\arguments{
\item{id}{{\link[base]{character}} expected. Vector containing all desired identifiers at once}

\item{result_check}{{\link[base]{character}} expected. Vector containing the identifiers to be verified}
}
\value{
The function returns a {\link[base]{character}}
}
\description{
Function to list elements with a number of occurrences other than 2
}
\examples{
text_object_more_or_less(id = c("1", "2", "3"), result_check = c("1", "3", "3"))
}
