% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_operation_inspector}
\alias{check_operation_inspector}
\title{Gives the inconsistencies between fishing success status, vessel activity, type of school or weight caught}
\usage{
check_operation_inspector(dataframe1, output)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_operation_inspector () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_operation_inspector function is to provide a table of data that contains an inconsistency with succes status and vessel activity, the type of school or the weight caught
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  activity_id}}
 \item{\code{  schooltype_code}}
 \item{\code{  successstatus_code}}
 \item{\code{  activity_weight}}
 \item{\code{  vesselactivity_code}}
}
}
