% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function.R
\name{check_weighting_inspector}
\alias{check_weighting_inspector}
\title{Gives the inconsistencies between the sample weighting and sample weight or landed weight}
\usage{
check_weighting_inspector(
  dataframe1,
  dataframe2,
  dataframe3,
  dataframe4,
  output,
  vessel_type = c("6", "2"),
  threshold_weight = 100,
  threshold_ratio = 0.95,
  sample_type_code_landing_baitboat = c("11"),
  landing_type_baitboat = c("L-YFT-10", "L-BET-10", "L-TUN-10"),
  threshold_baitboat = 1
)
}
\arguments{
\item{dataframe1}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_weighting_inspector () function.}

\item{dataframe2}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_weighting_inspector () function.}

\item{dataframe3}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_weighting_inspector () function.}

\item{dataframe4}{{\link[base]{data.frame}} expected. Csv or output of the function {\link[furdeb]{data_extraction}}, which must be done before using the check_weighting_inspector () function.}

\item{output}{{\link[base]{character}} expected. Kind of expected output. You can choose between "message", "report" or "logical".}

\item{vessel_type}{{\link[base]{character}} expected. Default values: c("6", "2"). List of two elements, the first is the seine vessel type code, and the second is the baitboat type code.}

\item{threshold_weight}{{\link[base]{numeric}} expected. Default values: 100. Seiner threshold weight}

\item{threshold_ratio}{{\link[base]{numeric}} expected. Default values: 0.95. Percentage threshold between weight and weighted weight for seiners}

\item{sample_type_code_landing_baitboat}{{\link[base]{character}} expected. Default values: c("11"). List of sample type codes for baitboat fresh landings}

\item{landing_type_baitboat}{{\link[base]{character}} expected. Default values: c("L-YFT-10", "L-BET-10", "L-TUN-10"). List of codes for fresh baitboat landings}

\item{threshold_baitboat}{{\link[base]{numeric}} expected. Default values: 1. Threshold for baitboats with respect to the difference between the weighted weight and the landed fresh weight and the difference between the weight and the weighted weight}
}
\value{
The function returns a {\link[base]{character}} with output is "message", a {\link[base]{data.frame}} with output is "report", a {\link[base]{logical}} with output is "logical"
}
\description{
The purpose of the check_weighting_inspector  function is to provide a table of data that contains an inconsistency between the sample weighting
}
\details{
The input dataframe must contain all these columns for the function to work :
\itemize{
Dataframe 1:
 \item{\code{  sample_id}}
 \item{\code{  sample_smallsweight}}
 \item{\code{  sample_bigsweight}}
 \item{\code{  sample_totalweight}}
 \item{\code{  trip_id}}
 \item{\code{  sampletype_code}}
}
\itemize{
Dataframe 2:
 \item{\code{  sample_id}}
 \item{\code{  sampleactivity_weightedweight}}
}
\itemize{
Dataframe 3:
 \item{\code{  trip_id}}
 \item{\code{  vesseltype_code}}
 \item{\code{  vesseltype_label1}}
}
\itemize{
Dataframe 4:
 \item{\code{  landing_id}}
 \item{\code{  landing_weight}}
 \item{\code{  weightcategory_code}}
 \item{\code{  trip_id}}
}
}
\examples{
dataframe1 <- data.frame(sample_id = c("1", "2", "3", "4", "5"),
                         sample_smallsweight = c(10, 32, 2.5, 30, 12),
                         sample_bigsweight = c(50, 2, 9, 3, 6),
                         sample_totalweight = c(NA, NA, NA, 33, 8),
                         trip_id =  c("1", "1", "2", "3", "4"),
                         sampletype_code = c("1", "1", "1", "11", "11"))
dataframe2 <- data.frame(sample_id = c("1", "1", "2", "3", "4", "5"),
                         sampleactivity_weightedweight = c(70, 5, 18, 12, 33, 5))
dataframe3 <- data.frame(trip_id = c("1", "2", "3", "4"),
          vesseltype_code = c("6", "6", "2", "2"),
          vesseltype_label1 = c("vessel_type_1", "vessel_type_1", "vessel_type_2", "vessel_type_2"))
dataframe4 <- data.frame(landing_id = c("1", "2", "3", "4", "5"),
                         landing_weight = c(85, 26, 30, 2.6, 20),
                         weightcategory_code = c("W-1", "W-1", "L-YFT-10", "L-YFT-10", "L-YFT-10"),
                         trip_id = c("1", "2", "3", "3", "4"))
check_weighting_inspector(dataframe1, dataframe2, dataframe3, dataframe4, output = "report")
}
