% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_tab_content.R
\name{mod_tab_content_server}
\alias{mod_tab_content_server}
\title{tab_content Server Functions}
\usage{
mod_tab_content_server(
  id,
  tab_info,
  check_info,
  type_check_info,
  config_data,
  res_auth
)
}
\arguments{
\item{id}{Internal parameters for shiny}

\item{tab_info}{{\link[base]{list}} expected. Information about the dynamic tab display}

\item{check_info}{{\link[base]{list}} expected. Information about the dynamic check display}

\item{type_check_info}{{\link[base]{list}} expected. Information about the dynamic type of check}

\item{config_data}{{\link[base]{list}} expected. Reactive value contains the contents of the configuration file}

\item{res_auth}{{\link[base]{character}} expected. Reactive value containing information about the user's connection}
}
\value{
The function returns nothing, instantiating the tab
}
\description{
A shiny Module for creation of all tab
}
\details{
\itemize{
tab_info:
 \item{\code{  id : Tab identification, mandatory, character expected}}
}
\itemize{
check_info:
 \item{\code{  id : Check identification, mandatory, character expected}}
 \item{\code{  tab : Tab identification, mandatory, character expected}}
 \item{\code{  title : Name to be displayed for check, optional, character expected}}
 \item{\code{  text : Text to be displayed for check, optional, character expected}}
 \item{\code{  type : Check type, you must choose between check type identifiers (type_check_info), mandatory, character expected}}
 \item{\code{  column_no_wrap : Column numbers that should not be subject to automatic line breaks, optional, integer expected}}
 \item{\code{  size_box : Check size specification for each window size type (format "col-sm-your_size col-md-your_size col-lg-your_size"), optional, character expected}}
}
\itemize{
type_check_info:
 \item{\code{  id : Button selection identification, mandatory, character expected}}
 \item{\code{  specific_check : Indicates whether the selection should display specific controls or not, optional (default : TRUE), logical expected}}
}
\itemize{
config_data:
 \item{\code{  databases_configuration : named list with database connection information}}
}
}
