% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_table.R
\name{mod_table_server}
\alias{mod_table_server}
\title{table Server Functions}
\usage{
mod_table_server(
  id,
  data,
  name,
  type_line_check,
  referential_file,
  column_no_wrap = NULL,
  name_function_plot = NULL,
  name_function_text_plot = NULL,
  title_window = NULL
)
}
\arguments{
\item{id}{Internal parameters for shiny}

\item{data}{{\link[base]{list}} expected. Reactive list containing results tables for all checks}

\item{name}{{\link[base]{character}} expected. Control identifier in data to be displayed}

\item{type_line_check}{{\link[base]{character}} expected. Reactive value containing the type of line selected by the user}

\item{referential_file}{{\link[base]{list}} expected. Reactive list containing referential tables for all plot}

\item{column_no_wrap}{{\link[base]{integer}} expected. Default values: NULL. Column numbers that should not be subject to automatic line breaks}

\item{name_function_plot}{{\link[base]{character}} expected. Default values: NULL. Name of the function that creates the plot}

\item{name_function_text_plot}{{\link[base]{character}} expected. Default values: NULL. Name of the function that creates the text to be displayed in the plot window}

\item{title_window}{{\link[base]{character}} expected. Default values: NULL. Plot window name}
}
\value{
The function returns nothing, instantiating the table
}
\description{
A shiny Module for creation of all table
}
