% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/start_configuration.R
\name{start_configuration}
\alias{start_configuration}
\title{Launch ui AkadoR}
\usage{
start_configuration(secure_connection = FALSE)
}
\arguments{
\item{secure_connection}{{\link[base]{logical}} expected. Default values: FALSE. If you wish to enable secure connection to the application, set TRUE}
}
\description{
The purpose of the start_configuration function is to configure startup, secure connection to the application, ...
}
